% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{averageClosenessCentrality}
\alias{averageClosenessCentrality}
\title{Average closeness centrality}
\usage{
averageClosenessCentrality(expr, labels, adjacencyMatrix)
}
\arguments{
\item{expr}{Matrix of genes (rows) vs microarrays (columns)}

\item{labels}{a vector in which a position indicates the phenotype of the corresponding sample or state}

\item{adjacencyMatrix}{a function that returns the adjacency matrix for a given variables values matrix}
}
\value{
a list of values containing the average closeness centrality of each network.
}
\description{
Average closeness centrality
}
