% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{averageClusteringCoefficient}
\alias{averageClusteringCoefficient}
\title{Average clustering coefficient}
\usage{
averageClusteringCoefficient(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Average clustering coefficient
}
