% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{averageDegreeCentrality}
\alias{averageDegreeCentrality}
\title{Average degree centrality}
\usage{
averageDegreeCentrality(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Average degree centrality
}
