% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{averageEigenvectorCentrality}
\alias{averageEigenvectorCentrality}
\title{Average eigenvector centrality}
\usage{
averageEigenvectorCentrality(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Average eigenvector centrality
}
