% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{averageShortestPath}
\alias{averageShortestPath}
\title{Average shortest path}
\usage{
averageShortestPath(expr, labels, adjacencyMatrix)
}
\arguments{
\item{expr}{Matrix of variables (columns) vs samples (rows)}

\item{labels}{a vector in which a position indicates the phenotype of the corresponding sample or state}

\item{adjacencyMatrix}{a function that returns the adjacency matrix for a given variables values matrix}
}
\value{
a list of values containing the average shortest path of each network.
}
\description{
Average shortest path
}
\examples{
set.seed(1)
expr <- as.data.frame(matrix(rnorm(120),40,30))
labels<-rep(0:3,10)
adjacencyMatrix1 <- adjacencyMatrix(method="spearman", association="pvalue",
 threshold="fdr", thr.value=0.05, weighted=FALSE)
averageShortestPath(expr, labels, adjacencyMatrix1)
}
