% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{betweennessCentralityTest}
\alias{betweennessCentralityTest}
\title{Betweenness centrality test}
\usage{
betweennessCentralityTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL)
}
\description{
Betweenness centrality test
}
