\name{bnsDataTest}
\docType{data}
\alias{bnsDataTest}
\title{Gene expression in gliomas}
\description{
  Expression of 134 target genes of NFKBIA in four gliomas tissues, 
astrocytom, oligodendrogliom, oligoastrcytom and Primary Multiform Glioblastom.
}
\usage{bnsDataTest}
\format{A data frame containing 134 variables (colunms) and 658 observations.}
\source{TCGA - The Cancer Genome Atlas}
\references{
  KINKER, G. S. et al. Deletion and low expression of NFKBIA are associated with poor prognosis in lower-grade glioma patients. Scientific Reports, v. 6, n. April, p. 24160, 2016.
}
\keyword{datasets}
