% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{closenessCentrality}
\alias{closenessCentrality}
\title{Closeness centrality}
\usage{
closenessCentrality(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Closeness centrality
}
