% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{closenessCentralityTest}
\alias{closenessCentralityTest}
\title{Closeness centrality test}
\usage{
closenessCentralityTest(expr, labels, adjacencyMatrix, numPermutations = 1000,
  options = NULL)
}
\description{
Closeness centrality test
}
