% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{closenessCentralityVertexTest}
\alias{closenessCentralityVertexTest}
\title{Closeness centrality test for vectors}
\usage{
closenessCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL)
}
\description{
Closeness centrality test for vectors
}
