% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{clusteringCoefficient}
\alias{clusteringCoefficient}
\title{Clustering coefficient}
\usage{
clusteringCoefficient(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Clustering coefficient
}
