% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{clusteringCoefficientTest}
\alias{clusteringCoefficientTest}
\title{Clustering coefficient test}
\usage{
clusteringCoefficientTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL)
}
\description{
Clustering coefficient test
}
