% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{degreeCentralityVertexTest}
\alias{degreeCentralityVertexTest}
\title{Degree centrality test for Vector}
\usage{
degreeCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL)
}
\description{
Degree centrality test for Vector
}
