% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{degreeDistributionTest}
\alias{degreeDistributionTest}
\title{Degree distribution test}
\usage{
degreeDistributionTest(expr, labels, adjacencyMatrix, numPermutations = 1000,
  options = list(bandwidth = "Sturges"))
}
\arguments{
\item{expr}{Matrix of genes (rows) vs microarrays (columns)}

\item{labels}{a vector in which a position indicates the phenotype of the corresponding sample or state}

\item{adjacencyMatrix}{a function that returns the adjacency matrix for a given variables values matrix}

\item{numPermutations}{number of permutations that will be carried out in the permutation test}

\item{options}{a list containing parameters. It can be set to either
\code{list(bandwidth="Sturges")} or \code{list(bandwidth="Silverman")}.}
}
\description{
Degree distribution test
}
