% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files_bns.R
\name{diffNetAnalysis}
\alias{diffNetAnalysis}
\title{Class vector of data table}
\usage{
diffNetAnalysis(method, options, varFile, labels, varSets = NULL,
  adjacencyMatrix, numPermutations = 1000, print = TRUE,
  resultsFile = NULL, seed = NULL, min.vert = 5)
}
\arguments{
\item{method}{a method receives two adjacency matrices and returns a list containing a statistic theta that measures the difference between them, and a p-value for the test H0: theta = 0 against H1: theta > 0.}

\item{options}{a list contaning paremeters used by 'method'.}

\item{varFile}{a numeric matrix contaning variables values data.}

\item{labels}{a vector of -1s, 0s, and 1s associating each sample with a phenotype. The value 0 corresponds to the first phenotype class of interest, 1 to the second phenotype class of interest, and -1 to the other classes, if there are more than two classes in the gene expression data.}

\item{varSets}{a list of gene sets. Each element of the list is a character vector v, where v[1] contains the gene set name, v[2] descriptions about the set, v[3..length(v)] the genes that belong to the set.}

\item{adjacencyMatrix}{a function that receives a numeric matrix containing gene expression data and returns the adjacency matrix of the inferred co-expression graph.}

\item{numPermutations}{the number of permutations for the permutation test.}

\item{print}{a logical. If true, it prints execution messages on the screen. resultsFile: path to a file where the partial results of the analysis will be saved. If NULL, then no partial results are saved.}

\item{seed}{the seed for the random number generators. If it is not null then the sample permutations are the same for all the gene sets.}
}
\value{
a data frame containing the name, size, test statistic, nominal p-value and adjusted p-value (q-value) associated with each gene set.
}
\description{
Class vector of data table
}
