% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{doLabels}
\alias{doLabels}
\title{Class vector of data table}
\usage{
doLabels(fileName, factorName = NULL, classes = NULL, dec = ".",
  sep = ";")
}
\arguments{
\item{fileName}{the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, getwd().}

\item{factorName}{string indicating the column name used to determine the labels of each row of matrix data. The NULL (default) indicates that the first column will be used.}

\item{classes}{a vector of strings indicating which labels of choosed column will be compared, the minimum are two labels. The NULL (default) indicates that all classes will be compared.}

\item{dec}{the character used in the file for decimal points.}

\item{sep}{the field separator character. Values on each line of the file are separated by this character. If sep = "" the separator is white space, that is one or more spaces, tabs, newlines or carriage returns, if sep=NULL (default), the function uses tabulation for .txt files or ";" for .csv files.}
}
\description{
Class vector of data table
}
