% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{eigenvectorCentrality}
\alias{eigenvectorCentrality}
\title{Eigenvector centrality}
\usage{
eigenvectorCentrality(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Eigenvector centrality
}
