% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{eigenvectorCentralityTest}
\alias{eigenvectorCentralityTest}
\title{Eigenvector centrality test}
\usage{
eigenvectorCentralityTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL)
}
\description{
Eigenvector centrality test
}
