% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{eigenvectorCentralityVertexTest}
\alias{eigenvectorCentralityVertexTest}
\title{Eigenvector centrality test for vectors}
\usage{
eigenvectorCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL)
}
\description{
Eigenvector centrality test for vectors
}
