% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{JS}
\alias{JS}
\title{Given two spectral densities, returns the Jensen-Shannon divergence between
the corresponding graphs}
\usage{
JS(f1, f2)
}
\arguments{
\item{f1}{an igraph graph object}

\item{f2}{an igraph graph object}
}
\value{
returns the Jensen-Shannon divergence between the corresponding graphs
}
\description{
'JS' computes the Jensen-Shannon divergence between the spectral density
functions of two graphs
}
\seealso{
\code{density}
}
