% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{JSdegree}
\alias{JSdegree}
\title{Jensen-Shannon divergence between the density functions of the degrees of
two graphs}
\usage{
JSdegree(G1, G2, options = list(bandwidth = "Sturges"))
}
\arguments{
\item{G1}{an igraph graph object}

\item{G2}{an igraph graph object}

\item{options}{a list containing parameters. It can be set to either 
\code{list(bandwidth="Sturges")} or \code{list(bandwidth="Silverman")}.}
}
\value{
a list containing the components, f1 (density estimate of the 
graph G1), and f2 (density estimate of the graph G2). Each component is
a list, where the first element is the vector 'x' of 'npoints' coordinates
of the points where the density function i estimated, and the second is
a vector 'y' of the estimated density values.
}
\description{
'JSdegree' computes the Jensen-Shannon divergence between the density 
functions of the degrees of two graphs
}
\seealso{
\code{graph.strength}

\code{density}
}
