% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{KLdegree}
\alias{KLdegree}
\title{Kullback-Liebler divergence among the density functions of the degrees of
two or more graphs}
\usage{
KLdegree(f)
}
\arguments{
\item{f}{a list containing the components 'x' and 'densities'. 
The first element is the vector 'x' of 'npoints' coordinates
of the points where the density function i estimated, and the second is
a vector 'y' of the estimated density values.}
}
\value{
returns a list containing the components 'theta' and 'partial'. 
'theta' is a value representaing the Kullback-Liebler divergence among the corresponding distributions.
'partial' is a vector of KL divergences between each network distribuiton and the average degree distribution.
}
\description{
'KLdegree' computes the Kullback-Liebler divergence among the density
functions of the degrees of two or more graphs
}
\examples{
G<-list()                                   
G[[1]]<-erdos.renyi.game(30,0.6)
G[[2]]<-barabasi.game(30,power = 1)
G[[3]]<-watts.strogatz.game(2,30,2,0.3)
f<-nDegreeDensities(G, npoints=1024, bandwidth="Sturges")
KLdegree(f)
}
\seealso{
\code{graph.strength}

\code{density}
}
