% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{KLspectrum}
\alias{KLspectrum}
\title{Kullback-Liebler divergence among the spectral density functions of 
two or more graphs}
\usage{
KLspectrum(f)
}
\arguments{
\item{f}{a list containing the components 'x' and 'densities'. 
The first element is the vector 'x' of 'npoints' coordinates
of the points where the density function i estimated, and the second is
a vector 'y' of the estimated density values.}
}
\value{
returns a list containing the components 'theta' and 'partial'. 
'theta' is a value representaing the Kullback-Liebler divergence among the corresponding distributions.
'partial' is a vector of KL divergences between each network distribuiton and the average spectral distribution.
}
\description{
'KLspectrum' computes the Kullback-Liebler divergence among the spectral density
functions of two or more graphs
}
\examples{
A<-list()
A[[1]]<-as.matrix(as_adj(erdos.renyi.game(30,0.6,directed = FALSE)))
A[[2]]<-as.matrix(as_adj(barabasi.game(30,power = 1,directed = FALSE)))
A[[3]]<-as.matrix(as_adj(watts.strogatz.game(1,30,2,0.3)))
f<-nSpectralDensities(A, bandwidth="Sturges")
KLspectrum(f)
}
\seealso{
\code{graph.strength}

\code{density}
}
