% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{labels}
\alias{labels}
\alias{varFile}
\title{Labels of glioma tissues in gene expression}
\format{Vector which associates each row of data.frame (varFile) with a state (that will be an network).}
\source{
\href{https://cancergenome.nih.gov/}{TCGA - The Cancer Genome Atlas}
}
\usage{
data(labels)
}
\description{
Labels of glioma tissues in gene expression
}
\examples{
data(labels)
# Run BNS analysis
}
\references{
KINKER, G. S. et al. Deletion and low expression of NFKBIA are associated with poor prognosis in lower-grade glioma patients. Scientific Reports, v. 6, n. April, p. 24160, 2016.
}
\keyword{datasets}
