% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{nodeTest}
\alias{nodeTest}
\alias{degreeCentralityVertexTest}
\alias{betweennessCentralityVertexTest}
\alias{closenessCentralityVertexTest}
\alias{eigenvectorCentralityVertexTest}
\alias{clusteringCoefficientVertexTest}
\title{Node score equality test}
\usage{
degreeCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL, BPPARAM = MulticoreParam())

betweennessCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL, BPPARAM = MulticoreParam())

closenessCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL, BPPARAM = MulticoreParam())

eigenvectorCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL, BPPARAM = MulticoreParam())

clusteringCoefficientVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = NULL, BPPARAM = MulticoreParam())
}
\arguments{
\item{expr}{Matrix of variables (columns) vs samples (rows)}

\item{labels}{a vector in which a position indicates the phenotype of the corresponding sample or state}

\item{adjacencyMatrix}{a function that returns the adjacency matrix for a given variables values matrix}

\item{numPermutations}{number of permutations that will be carried out in the permutation test}

\item{options}{argument non used in this function}

\item{BPPARAM}{An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation, or a list of BiocParallelParam instances, to be applied in sequence for nested calls to BiocParallel functions.}
}
\value{
A table, containing on the columns, the following informations for each variable (rows):
"Test Statistic" - difference among the degree centrality of a node in two or more networks associated with each phenotype
"Nominal p-value" - the Nominal p-value of the test
"Q-value" - the q-value of the test, correction of p-value by FDR to many tests
"Factor n" - the node degree centrality in each network compared
}
\description{
Nodes scores equality test between network
}
\examples{
set.seed(1)
expr <- as.data.frame(matrix(rnorm(120),40,30))
labels<-rep(0:3,10)
adjacencyMatrix1 <- adjacencyMatrix(method="spearman", association="pvalue",
 threshold="fdr", thr.value=0.05, weighted=FALSE)

# Degree centrality test
degreeCentralityVertexTest(expr, labels, adjacencyMatrix1,numPermutations=10)

# Betweenness centrality test
betweennessCentralityVertexTest(expr, labels, adjacencyMatrix1,numPermutations=10)

# Closeness centrality test
closenessCentralityVertexTest(expr, labels, adjacencyMatrix1,numPermutations=10)

# Eigenvector centrality test
eigenvectorCentralityVertexTest(expr, labels, adjacencyMatrix1,numPermutations=10)

# Clustering coefficient test
clusteringCoefficientVertexTest(expr, labels, adjacencyMatrix1,numPermutations=10)
}
