% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{nSpectralDensities}
\alias{nSpectralDensities}
\title{Spectral Density functions of n graphs}
\usage{
nSpectralDensities(A, from = NULL, to = NULL, bandwidth = "Silverman")
}
\arguments{
\item{A}{a list of adjacency matrices}

\item{from}{the lower value used to build the distribution}

\item{to}{the higher value used to build the distribution}

\item{bandwidth}{a parameters. It can be set to either "Sturges" or "Silverman".}
}
\value{
a list containing the components 'x' and 'densities'. 
The first element is the vector 'x' of 'npoints' coordinates
of the points where the density function i estimated, and the second is
a vector 'y' of the estimated density values.
}
\description{
Returns the spectral densities for a list of adjacency matrices at the
same points
}
\examples{
A<-list()
A[[1]]<-as.matrix(as_adj(erdos.renyi.game(30,0.6,directed = FALSE)))
A[[2]]<-as.matrix(as_adj(barabasi.game(30,power = 1,directed = FALSE)))
A[[3]]<-as.matrix(as_adj(watts.strogatz.game(1,30,2,0.3)))
d<-nSpectralDensities(A, bandwidth="Sturges")
par(mfrow=c(1,3))
plot(d$x,d$densities[,1],main="Erdos-Renyi\\n Spectral distribution",
xlab="Eigenvalue",ylab="Frequency")
plot(d$x,d$densities[,2],main="Barabasi\\n Spectral distribution",
xlab="Eigenvalue",ylab="Frequency")
plot(d$x,d$densities[,3],main="Watts-Strogatz\\n Spectral distribution",
xlab="Eigenvalue",ylab="Frequency")
}
\seealso{
\code{KLdegree}

\code{density}
}
