% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{readSetFile}
\alias{readSetFile}
\title{Read a collection of variables sets (*.txt)}
\usage{
readSetFile(fileName)
}
\arguments{
\item{fileName}{a string containing the file name}
}
\value{
a list of gene sets. Each element of the list is a character vector
v, where v[1] contains the gene set name, v[2] descriptions about the set,
v[3..length(v)] the genes that belong to the set.
}
\description{
'readSetFile' reads a tab-delimited text file containing a collection of
gene sets.
}
\examples{
# Read example set file
set_fname <- system.file("extdata", "variableSet_BioNetStat_tutorial_data.gmt", package = "BioNetStat")
deneSets <- readSetFile(set_fname)
}
