% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{shortestPathTest}
\alias{shortestPathTest}
\title{Shortest path test}
\usage{
shortestPathTest(expr, labels, adjacencyMatrix, numPermutations = 1000,
  options = NULL)
}
\arguments{
\item{expr}{Matrix of variables (columns) vs samples (rows)}

\item{labels}{a vector in which a position indicates the phenotype of the corresponding sample or state}

\item{adjacencyMatrix}{a function that returns the adjacency matrix for a given variables values matrix}

\item{numPermutations}{number of permutations that will be carried out in the permutation test}

\item{options}{argument non used in this function}
}
\value{
A list containing:
"measure" - difference among the shortests paths of two or more networks associated with each phenotype
"p.value" - the Nominal p-value of the test for the Jensen-Shannon divergence
"Partial" - a vector with the weigths of each network in a measure value
}
\description{
Shortest path test
}
\examples{
set.seed(1)
expr <- as.data.frame(matrix(rnorm(120),40,30))
labels<-rep(0:3,10)
adjacencyMatrix1 <- adjacencyMatrix(method="spearman", association="pvalue",
 threshold="fdr", thr.value=0.05, weighted=FALSE)
shortestPathTest(expr, labels, adjacencyMatrix1,numPermutations=10)
}
