% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{spectralDistributionTest}
\alias{spectralDistributionTest}
\title{Spectral distribution test}
\usage{
spectralDistributionTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000, options = list(bandwidth = "Sturges"))
}
\description{
Spectral distribution test
}
