% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{spectralEntropies}
\alias{spectralEntropies}
\title{Spectral entropies}
\usage{
spectralEntropies(expr, labels, adjacencyMatrix, options = list(bandwidth =
  "Sturges"))
}
\arguments{
\item{expr}{Matrix of variables (columns) vs samples (rows)}

\item{labels}{a vector in which a position indicates the phenotype of the corresponding sample or state}

\item{adjacencyMatrix}{a function that returns the adjacency matrix for a given variables values matrix}

\item{options}{a list containing parameters. It can be set to either
\code{list(bandwidth="Sturges")} or \code{list(bandwidth="Silverman")}.}
}
\value{
a list of values containing the spectral entropy of each network.
}
\description{
Spectral entropies
}
\examples{
set.seed(1)
expr <- as.data.frame(matrix(rnorm(120),40,30))
labels<-rep(0:3,10)
adjacencyMatrix1 <- adjacencyMatrix(method="spearman", association="pvalue",
 threshold="fdr", thr.value=0.05, weighted=FALSE)
spectralEntropies(expr, labels, adjacencyMatrix1, options=list(bandwidth="Sturges"))
}
