% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{varFile}
\alias{varFile}
\title{Gene expression in glioma tissues}
\format{Object of class \code{data.frame} containing 134 variables (colunms) and 658 observations.}
\source{
\href{https://cancergenome.nih.gov/}{TCGA - The Cancer Genome Atlas}
}
\usage{
data(varFile)
}
\description{
Gene expression in glioma tissues
}
\examples{
data(varFile)
# Run BNS analysis
}
\references{
KINKER, G. S. et al. Deletion and low expression of NFKBIA are associated with poor prognosis in lower-grade glioma patients. Scientific Reports, v. 6, n. April, p. 24160, 2016.
}
\keyword{datasets}
