% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{averageBetweennessCentrality}
\alias{averageBetweennessCentrality}
\title{Average betweenness centrality}
\usage{
averageBetweennessCentrality(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Average betweenness centrality
}
