% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{averageClosenessCentrality}
\alias{averageClosenessCentrality}
\title{Average closeness centrality}
\usage{
averageClosenessCentrality(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Average closeness centrality
}
