% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{averageShortestPath}
\alias{averageShortestPath}
\title{Average shortest path}
\usage{
averageShortestPath(expr, labels, adjacencyMatrix, options = NULL)
}
\description{
Average shortest path
}
