% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{degreeDensities}
\alias{degreeDensities}
\title{Density functions of the degrees of two graphs}
\usage{
degreeDensities(G1, G2, npoints = 1024, options = list(bandwidth =
  "Sturges"))
}
\arguments{
\item{G1}{an igraph graph object}

\item{G2}{an igraph graph object}

\item{npoints}{number of points used in density function estimation}

\item{options}{a list containing parameters. It can be set to either
\code{list(bandwidth="Sturges")} or \code{list(bandwidth="Silverman")}.}
}
\value{
a list containing the components, f1 (density estimate of the
graph G1), and f2 (density estimate of the graph G2). Each component is
a list, where the first element is the vector 'x' of 'npoints' coordinates
of the points where the density function is estimated, and the second is
a vector 'y' of the estimated density values.
}
\description{
'degreeDensities' estimates the density functions of the degrees for two
graphs at the same coordinates
}
\examples{
G1<-erdos.renyi.game(30,0.6)
G2<-barabasi.game(30,power = 1)
d<-degreeDensities(G1, G2, npoints=1024, options=list(bandwidth="Sturges"))
par(mfrow=c(1,2))
plot(d$f1$x,d$f1$y,main="Erdos-Renyi\\n Degree distribution",
xlab="Degree",ylab="Frequency")
plot(d$f2$x,d$f2$y,main="Barabasi\\n Degree distribution",
xlab="Degree",ylab="Frequency")
}
\seealso{
\code{graph.strength}

\code{density}
}
