% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{degreeDistributionTest}
\alias{degreeDistributionTest}
\title{Degree distribution test}
\usage{
degreeDistributionTest(expr, labels, adjacencyMatrix, numPermutations = 1000,
  options = list(bandwidth = "Sturges"))
}
\description{
Degree distribution test
}
