% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature.R
\name{eigenvectorCentralityVertexTest}
\alias{eigenvectorCentralityVertexTest}
\title{Eigenvector centrality test for vectors}
\usage{
eigenvectorCentralityVertexTest(expr, labels, adjacencyMatrix,
  numPermutations = 1000)
}
\arguments{
\item{expr}{Matrix of genes (rows) vs microarrays (columns)}

\item{labels}{a vector in which a position indicates the phenotype of the corresponding sample or state}

\item{adjacencyMatrix}{a function that returns the adjacency matrix for a given variables values matrix}

\item{numPermutations}{number of permutations that will be carried out in the permutation test}
}
\description{
Eigenvector centrality test for vectors
}
