% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{nDegreeDensities}
\alias{nDegreeDensities}
\title{Density functions of the degrees of two graphs}
\usage{
nDegreeDensities(Gs, npoints = 1024, bandwidth = "Sturges", from = NULL,
  to = NULL)
}
\arguments{
\item{G1}{an igraph graph object}

\item{G2}{an igraph graph object}
}
\value{
a list containing the components, f1 (density estimate of the 
graph G1), and f2 (density estimate of the graph G2). Each component is
a list, where the first element is the vector 'x' of 'npoints' coordinates
of the points where the density function is estimated, and the second is 
a vector 'y' of the estimated density values.
}
\description{
'degreeDensities' estimates the density functions of the degrees for two 
graphs at the same coordinates
}
\seealso{
\code{graph.strength}

\code{density}
}
