% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{readVarFile}
\alias{readVarFile}
\title{Read variable values matrix}
\usage{
readVarFile(fileName, path = NULL, dec = ".", sep = NULL,
  check.names = T)
}
\arguments{
\item{fileName}{the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, getwd().}

\item{path}{the path to the directory that contains the file}

\item{dec}{the character used in the file for decimal points.}

\item{sep}{the field separator character. Values on each line of the file are separated by this character. If sep = "" the separator is white space, that is one or more spaces, tabs, newlines or carriage returns, if sep=NULL (default), the function uses tabulation for .txt files or ";" for .csv files.}

\item{check.names}{a logical value. If TRUE, the names of the data table kept as they are. Otherwise, the blank space, "-","/" and ",", are replaced by dots.}
}
\description{
Read variable values matrix
}
