% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{shortestPathTest}
\alias{shortestPathTest}
\title{Shortest path test}
\usage{
shortestPathTest(expr, labels, adjacencyMatrix, numPermutations = 1000,
  options = NULL)
}
\description{
Shortest path test
}
