% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_feature_vini_bns_rede_media.R
\name{spectralEntropies}
\alias{spectralEntropies}
\title{Spectral entropies}
\usage{
spectralEntropies(expr, labels, adjacencyMatrix, options = list(bandwidth =
  "Sturges"))
}
\description{
Spectral entropies
}
